+++ flutter_secure_storage/android/src/main/java/com/it_nomads/fluttersecurestorage/FlutterSecureStoragePlugin.java
@@ -1,5 +1,6 @@
 package com.it_nomads.fluttersecurestorage;
 
+import android.content.pm.PackageManager;
 import android.os.Handler;
 import android.os.HandlerThread;
 import android.os.Looper;
@@ -24,6 +25,7 @@ public class FlutterSecureStoragePlugin implements MethodCallHandler, FlutterPlu
     private HandlerThread workerThread;
     private Handler workerThreadHandler;
     private FlutterPluginBinding binding;
+    private boolean isStrongBoxAvailable;
 
     @Override
     public void onAttachedToEngine(FlutterPluginBinding binding) {
@@ -52,6 +54,7 @@ public class FlutterSecureStoragePlugin implements MethodCallHandler, FlutterPlu
         if (secureStorage != null) return true;
 
         try {
+            isStrongBoxAvailable = binding.getApplicationContext().getPackageManager().hasSystemFeature(PackageManager.FEATURE_STRONGBOX_KEYSTORE);
             secureStorage = new FlutterSecureStorage(binding.getApplicationContext(), options);
             return true;
         } catch (Exception e) {
@@ -123,6 +126,9 @@ public class FlutterSecureStoragePlugin implements MethodCallHandler, FlutterPlu
                 case "deleteAll":
                     handleDeleteAll(result);
                     break;
+                case "isStrongBoxSupported":
+                    handleStrongBoxAvailable(result);
+                    break;
                 default:
                     result.notImplemented();
             }
@@ -162,6 +168,10 @@ public class FlutterSecureStoragePlugin implements MethodCallHandler, FlutterPlu
         private void handleDeleteAll(Result result) {
             secureStorage.deleteAll();
             result.success(null);
+        }
+
+        private void handleStrongBoxAvailable(Result result) {
+            result.success(isStrongBoxAvailable);
         }
 
         @SuppressWarnings("unchecked")