+++ flutter_secure_storage_platform_interface/lib/flutter_secure_storage_platform_interface.dart
@@ -120,4 +120,17 @@ abstract class FlutterSecureStoragePlatform extends PlatformInterface {
   Future<void> deleteAll({
     required Map<String, String> options,
   });
+
+  /// Checks if the android device supports secure hardware-backed storage.
+  ///
+  /// Returns:
+  /// - A [Future] that resolves to `true` if the device supports secure
+  /// hardware-backed storage, or `false` otherwise.
+  Future<bool> isStrongBoxSupported({
+    required Map<String, String> options,
+  }) {
+    throw UnsupportedError(
+      'isStrongBoxSupported() is not available on this platform',
+    );
+  }
 }